requirements.txt
Django==4.1.3
djangorestframework==3.14.0
python-dotenv==1.0.0
djongo==1.3.6
firebase-admin==6.2.0
pymongo==3.12.3
httpcore==0.15.0
httpx==0.25.1
requests==2.28.1